feat: add endpoint of update software engineer by id

+++ src/main/java/com/amigoscode/SoftwareEngineerService.java
@@ -26,6 +26,21 @@ public class SoftwareEngineerService {
     }
 
     public void deleteSoftwareEngineerById(Integer id) {
+        boolean exists = softwareEngineerRepository.existsById(id);
+        if (!exists) {
+            throw new IllegalStateException(
+                    id + " not found"
+            );
+        }
         softwareEngineerRepository.deleteById(id);
+    }
+
+    public void updateSoftwareEngineerById(Integer id, SoftwareEngineer oldSoftwareEngineer) {
+        SoftwareEngineer newSoftwareEngineer = softwareEngineerRepository.findById(id)
+                .orElseThrow(() -> new IllegalStateException(
+                        "id value: " +id + " not found"));
+        newSoftwareEngineer.setName(oldSoftwareEngineer.getName());
+        newSoftwareEngineer.setTechStack(oldSoftwareEngineer.getTechStack());
+        softwareEngineerRepository.save(newSoftwareEngineer);
     }
 }
